refactor and cleanup after feedback
